refactor: separate potionEffects.test.ts into one file per potion type
